Fixed integration test failure
Integration tests failed as size() function was changed from long to Long (https://github.com/FileAnalysisSuite/adaptersdk-impl-jaxrs/pull/9). Convert to primitive so that the assert statement works.

+++ src/test/java/io/github/fileanalysissuite/adaptersdk/impls/jaxrs/AdapterSDKFakeTest.java
@@ -97,7 +97,7 @@ final class AdapterSDKFakeTest extends JerseyTest
         assertThat(failureDetails, equalTo(actualRetrieveFileListResponse.getFailures()));
         final List<FileListItem> files = actualRetrieveFileListResponse.getFiles();
         final FileListItem item = files.get(0);
-        assertThat(files.size(), equalTo(1));
+        assertThat(((long) files.size()), equalTo(1L));
         assertThat(item.getPartitionHint(), equalTo("-"));
         assertThat(item.getFileMetadata(), equalTo(fileMetadata));
     }
